requirements.txt
pillow
python-dotenv
typing